Move mappings of assets to custom files
