+++ rotkehlchen/assets/exchanges_mappings/__init__.py
@@ -1,0 +1,35 @@
+from rotkehlchen.assets.exchanges_mappings.binance import WORLD_TO_BINANCE  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.bitfinex import WORLD_TO_BITFINEX  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.bitpanda import WORLD_TO_BITPANDA  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.bitstamp import WORLD_TO_BITSTAMP  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.bittrex import WORLD_TO_BITTREX
+from rotkehlchen.assets.exchanges_mappings.coinbase import WORLD_TO_COINBASE  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.coinbase_pro import WORLD_TO_COINBASE_PRO  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.cryptocom import WORLD_TO_CRYPTOCOM  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.ftx import WORLD_TO_FTX  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.gemeni import WORLD_TO_GEMINI  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.iconomi import WORLD_TO_ICONOMI  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.kraken import WORLD_TO_KRAKEN  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.kucoin import WORLD_TO_KUCOIN  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.nexo import WORLD_TO_NEXO  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.poloniex import WORLD_TO_POLONIEX  # noqa: F401
+from rotkehlchen.assets.exchanges_mappings.uphold import WORLD_TO_UPHOLD  # noqa: F401
+
+__all__ = [
+    'WORLD_TO_BINANCE',
+    'WORLD_TO_BITFINEX',
+    'WORLD_TO_BITPANDA',
+    'WORLD_TO_BITSTAMP',
+    'WORLD_TO_BITTREX',
+    'WORLD_TO_COINBASE_PRO',
+    'WORLD_TO_COINBASE',
+    'WORLD_TO_CRYPTOCOM',
+    'WORLD_TO_FTX',
+    'WORLD_TO_GEMINI',
+    'WORLD_TO_ICONOMI',
+    'WORLD_TO_KUCOIN',
+    'WORLD_TO_KRAKEN',
+    'WORLD_TO_NEXO',
+    'WORLD_TO_POLONIEX',
+    'WORLD_TO_UPHOLD',
+]
