+++ rotkehlchen/assets/exchanges_mappings/cryptocom.py
@@ -1,0 +1,6 @@
+from rotkehlchen.constants.resolver import strethaddress_to_identifier
+
+WORLD_TO_CRYPTOCOM = {
+    strethaddress_to_identifier('0xBB0E17EF65F82Ab018d8EDd776e8DD940327B28b'): 'AXS',
+    'LUNA-2': 'LUNA',
+}
