+++ rotkehlchen/assets/exchanges_mappings/nexo.py
@@ -1,0 +1,7 @@
+from rotkehlchen.constants.resolver import strethaddress_to_identifier
+
+WORLD_TO_NEXO = {
+    strethaddress_to_identifier('0xB62132e35a6c13ee1EE0f84dC5d40bad8d815206'): 'NEXONEXO',
+    'GBP': 'GBPX',
+    strethaddress_to_identifier('0xdAC17F958D2ee523a2206206994597C13D831ec7'): 'USDTERC',
+}
